+++ __tests__/entity-creation.unit.test.js
@@ -1,6 +1,6 @@
 // Require Table and Entity classes
-const Table = require('../classes/Table')
-const Entity = require('../classes/Entity')
+const Table = require('../src/classes/Table')
+const Entity = require('../src/classes/Entity')
 
 const { DocumentClient } = require('./bootstrap-tests')
