+++ __tests__/expressionBuilder.unit.test.js
@@ -1,8 +1,8 @@
-const expressionBuilder = require('../lib/expressionBuilder')
+const expressionBuilder = require('../src/lib/expressionBuilder')
 
 // Require Table and Entity classes
-const Table = require('../classes/Table')
-const Entity = require('../classes/Entity')
+const Table = require('../src/classes/Table')
+const Entity = require('../src/classes/Entity')
 
 // Create basic entity
 const TestEntity = new Entity({
