+++ __tests__/formatItem.unit.test.js
@@ -1,10 +1,10 @@
-const formatItem = require('../lib/formatItem')
+const formatItem = require('../src/lib/formatItem')
 
 const { DocumentClient } = require('./bootstrap-tests')
 
 // Require Table and Entity classes
-const Table = require('../classes/Table')
-const Entity = require('../classes/Entity')
+const Table = require('../src/classes/Table')
+const Entity = require('../src/classes/Entity')
 
 
 // Create basic table
