+++ __tests__/misc-tests.js
@@ -25,8 +25,8 @@ afterAll(async () => {
 
 
 // Require Table and Entity classes
-const Table = require('../classes/Table')
-const Entity = require('../classes/Entity')
+const Table = require('../src/classes/Table')
+const Entity = require('../src/classes/Entity')
 
 
 describe('Table creation', ()=> {
