+++ __tests__/projectionBuilder.unit.test.js
@@ -1,8 +1,8 @@
-const projectionBuilder = require('../lib/projectionBuilder')
+const projectionBuilder = require('../src/lib/projectionBuilder')
 
 // Require Table and Entity classes
-const Table = require('../classes/Table')
-const Entity = require('../classes/Entity')
+const Table = require('../src/classes/Table')
+const Entity = require('../src/classes/Entity')
 
 
 // Create basic table
