+++ __tests__/table-creation.unit.test.js
@@ -2,8 +2,8 @@
 const { DocumentClient } = require('./bootstrap-tests')
 
 // Require Table and Entity classes
-const Table = require('../classes/Table')
-const Entity = require('../classes/Entity')
+const Table = require('../src/classes/Table')
+const Entity = require('../src/classes/Entity')
 
 
 describe('Table creation', ()=> {
