+++ src/classes/Entity.js
