+++ src/lib/checkAttribute.js
