+++ src/lib/formatItem.js
