+++ src/lib/normalizeData.js
