+++ src/lib/parseCompositeKey.js
