+++ src/lib/parseEntity.js
