+++ src/lib/parseEntityAttributes.js
