+++ src/lib/parseMapping.js
