+++ src/lib/parseTable.js
