+++ src/lib/parseTableAttributes.js
