+++ src/lib/projectionBuilder.js
